Separate backend and npx
