+++ backend/package.json
@@ -21,6 +21,7 @@
     "cors": "^2.8.5",
     "express": "^4.17.1",
     "https": "^1.0.0",
+    "json-url": "^2.6.0",
     "open": "^7.4.2",
     "webpack-merge": "^5.7.3"
   },
