+++ package.json
@@ -58,7 +58,6 @@
     "fundamental-styles": "^0.5.0",
     "graphql": "^14.5.4",
     "graphql-tag": "^2.10.1",
-    "json-url": "^2.6.0",
     "lodash": "^4.17.19",
     "lodash.clonedeep": "^4.5.0",
     "merge": "2.1.0",
